onyx-bot/requirements.txt
python-telegram-bot==21.5
anthropic>=0.40.0
openai>=1.0.0
httpx>=0.27.0
requests==2.32.3
google-api-python-client==2.143.0
google-auth==2.34.0
